Rename package to alpakka
