+++ core/src/main/resources/reference.conf
@@ -27,7 +27,7 @@ cassandra-journal {
   # Will be ignored if the contact point list is defined by host:port pairs.
   port = 9042
 
-  # The implementation of akka.cassandra.session.SessionProvider
+  # The implementation of akka.stream.alpakka.cassandra.SessionProvider
   # is used for creating the Cassandra Session. By default the 
   # the ConfigSessionProvider is building the Cluster from configuration properties
   # but it is possible to replace the implementation of the SessionProvider
@@ -38,7 +38,7 @@ cassandra-journal {
   # ConfigSessionProvider and overriding the lookupContactPoints method.
   # It may optionally have a constructor with an ActorSystem and Config parameter.
   # The config parameter is this config section of the plugin.
-  session-provider = akka.cassandra.session.ConfigSessionProvider
+  session-provider = akka.stream.alpakka.cassandra.ConfigSessionProvider
 
   # The identifier that will be passed as parameter to the
   # ConfigSessionProvider.lookupContactPoints method. 
@@ -379,7 +379,7 @@ cassandra-snapshot-store {
   # Will be ignored if the contact point list is defined by host:port pairs.
   port = 9042
 
-  # The implementation of akka.cassandra.session.SessionProvider
+  # The implementation of akka.stream.alpakka.cassandra.SessionProvider
   # is used for creating the Cassandra Session. By default the
   # the ConfigSessionProvider is building the Cluster from configuration properties
   # but it is possible to replace the implementation of the SessionProvider
@@ -390,7 +390,7 @@ cassandra-snapshot-store {
   # ConfigSessionProvider and overriding the lookupContactPoints method.
   # It may optionally have a constructor with an ActorSystem and Config parameter.
   # The config parameter is this config section of the plugin.
-  session-provider = akka.cassandra.session.ConfigSessionProvider
+  session-provider = akka.stream.alpakka.cassandra.ConfigSessionProvider
 
   # The identifier that will be passed as parameter to the
   # ConfigSessionProvider.lookupContactPoints method.
