+++ core/src/main/scala/akka/persistence/cassandra/journal/CassandraRecovery.scala
@@ -15,7 +15,8 @@ import akka.persistence.cassandra.query.EventsByPersistenceIdStage.{ Extractors,
 import akka.stream.scaladsl.{ Sink, Source }
 import akka.util.OptionVal
 import scala.concurrent._
-import akka.cassandra.session._
+
+import akka.persistence.cassandra._
 
 trait CassandraRecovery extends CassandraTagRecovery with TaggedPreparedStatements {
   this: CassandraJournal =>
