+++ core/src/main/scala/akka/persistence/cassandra/query/EventsByTagStage.scala
@@ -13,7 +13,6 @@ import akka.persistence.cassandra.journal.{ BucketSize, TimeBucket }
 import akka.persistence.cassandra.query.EventsByTagStage._
 import akka.stream.stage.{ GraphStage, _ }
 import akka.stream.{ ActorMaterializer, Attributes, Outlet, SourceShape }
-import akka.cassandra.session._
 import akka.util.PrettyDuration._
 
 import scala.annotation.tailrec
