+++ core/src/main/scala/akka/persistence/cassandra/query/QueryActorPublisher.scala
@@ -14,8 +14,8 @@ import akka.pattern.pipe
 import akka.stream.actor.ActorPublisher
 import akka.stream.actor.ActorPublisherMessage.{ Cancel, Request, SubscriptionTimeoutExceeded }
 import com.datastax.driver.core.{ ResultSet, Row }
-import akka.cassandra.session._
 import akka.persistence.cassandra.query.QueryActorPublisher._
+import akka.persistence.cassandra.ListenableFutureConverter
 
 import scala.util.control.NonFatal
 import akka.annotation.InternalApi
