+++ core/src/main/scala/akka/persistence/cassandra/snapshot/CassandraSnapshotCleanup.scala
@@ -7,9 +7,9 @@ package akka.persistence.cassandra.snapshot
 import java.lang.{ Long => JLong }
 
 import akka.Done
-import akka.cassandra.session.scaladsl.CassandraSession
 import akka.persistence.SnapshotMetadata
 import akka.persistence.cassandra.journal.FixedRetryPolicy
+import akka.stream.alpakka.cassandra.scaladsl.CassandraSession
 import com.datastax.driver.core.policies.LoggingRetryPolicy
 
 import scala.concurrent.{ ExecutionContext, Future }
