+++ core/src/test/scala/akka/persistence/cassandra/query/scaladsl/CassandraReadJournalSpec.scala
@@ -4,11 +4,11 @@
 
 package akka.persistence.cassandra.query.scaladsl
 
-import akka.cassandra.session.CassandraMetricsRegistry
 import akka.persistence.cassandra.query.TestActor
 import akka.persistence.cassandra.{ CassandraLifecycle, CassandraSpec }
 import akka.persistence.journal.{ Tagged, WriteEventAdapter }
 import akka.persistence.query.NoOffset
+import akka.stream.alpakka.cassandra.CassandraMetricsRegistry
 import akka.stream.testkit.scaladsl.TestSink
 import com.typesafe.config.ConfigFactory
