+++ core/src/test/scala/akka/persistence/cassandra/snapshot/CassandraSnapshotCleanupSpec.scala
@@ -7,12 +7,12 @@ package akka.persistence.cassandra.snapshot
 import akka.Done
 import akka.actor.Props
 import akka.event.Logging
-import akka.cassandra.session.scaladsl.CassandraSession
 import akka.persistence.cassandra.{ CassandraSpec, Persister }
 import com.typesafe.config.ConfigFactory
 
 import scala.concurrent.{ ExecutionContext, Future }
 import Persister._
+import akka.stream.alpakka.cassandra.scaladsl.CassandraSession
 
 object CassandraSnapshotCleanupSpec {
   val config = ConfigFactory.parseString("""
