--- session/src/main/scala/akka/cassandra/session/impl/StorePathPasswordConfig.scala
@@ -1,8 +1,0 @@
-package akka.cassandra.session.impl
-
-import akka.annotation.InternalApi
-
-/**
-  * INTERNAL API
-  */
-@InternalApi private[session] case class StorePathPasswordConfig(path: String, password: String)
