+++ session/src/main/scala/akka/stream/alpakka/cassandra/javadsl/CassandraSession.scala
@@ -2,30 +2,25 @@
  * Copyright (C) 2016-2017 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.cassandra.session.javadsl
+package akka.stream.alpakka.cassandra.javadsl
 
-import java.util.{ List => JList }
-import java.util.Optional
 import java.util.concurrent.CompletionStage
 import java.util.function.{ Function => JFunction }
+import java.util.{ Optional, List => JList }
+
+import akka.{ Done, NotUsed }
+import akka.actor.ActorSystem
+import akka.event.LoggingAdapter
+import akka.stream.alpakka.cassandra.{ CassandraSessionSettings, SessionProvider }
+import akka.stream.alpakka.cassandra.scaladsl
+import akka.stream.javadsl.Source
+import com.datastax.driver.core._
 
 import scala.annotation.varargs
 import scala.collection.JavaConverters._
-import scala.compat.java8.FutureConverters._
 import scala.compat.java8.OptionConverters._
+import scala.compat.java8.FutureConverters._
 import scala.concurrent.ExecutionContext
-
-import akka.Done
-import akka.NotUsed
-import akka.actor.ActorSystem
-import akka.cassandra.session.{ CassandraSessionSettings, SessionProvider }
-import akka.event.LoggingAdapter
-import akka.stream.javadsl.Source
-import com.datastax.driver.core.BatchStatement
-import com.datastax.driver.core.PreparedStatement
-import com.datastax.driver.core.Row
-import com.datastax.driver.core.Session
-import com.datastax.driver.core.Statement
 
 /**
  * Data Access Object for Cassandra. The statements are expressed in
@@ -37,7 +32,7 @@ import com.datastax.driver.core.Statement
  *
  * All methods are non-blocking.
  */
-final class CassandraSession(delegate: akka.cassandra.session.scaladsl.CassandraSession) {
+final class CassandraSession(delegate: scaladsl.CassandraSession) {
 
   /**
    * Use this constructor if you want to create a stand-alone `CassandraSession`.
@@ -51,7 +46,7 @@ final class CassandraSession(delegate: akka.cassandra.session.scaladsl.Cassandra
       metricsCategory: String,
       init: JFunction[Session, CompletionStage[Done]]) =
     this(
-      new akka.cassandra.session.scaladsl.CassandraSession(
+      new scaladsl.CassandraSession(
         system,
         sessionProvider,
         settings,
